scripts: mx8_install_android: Burn bootloader0 to mmc boot0 partition
When installing Android from recovery SD card it is better to write bootloader0
in mmc boot0 partition (i.e for imx8mp mmcblk2boot0) and then to enable this
partition for boot process to start from.
This way it is better to keep the same approach for mmc partconf in the uuu
command list file.

Note: For some SoCs the bootloader0 offset in the eMMC boot0 partition is
different from bootloader0 offset in the user partition.

Signed-off-by: Natalia Kovalenko <natasha.k@variscite.com>