imx8mp-var-dart: Enable support for Secondary display with its own content
By default, the secondary display in 'Mirror Mode',
to have it in 'Extended Mode' set PRODUCT_MULTI_DISPLAY to true

The secondary display will use the system's Home Screen Launcher.
It is possible to use an alternative Launcher, by defining it in
config_secondaryHomePackage and then by setting
config_useSystemProvidedLauncherForSecondary to false.
To have the mouse pointer associated with this display we still need
to enable "Force Desktop Mode" in Developer Options or to issue this command:
'settings put global force_desktop_mode_on_external_displays 1'
and then reboot the device.

Known issue: The navigation bar HOME and BACK button icons are invisible.
There are still button mouse areas that receive mouse events in the left
corner at the bottom of the screen.

Signed-off-by: Natalia Kovalenko <natasha.k@variscite.com>